requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Computer Vision & ML
deepface==0.0.84
opencv-python==4.8.0.76
numpy==1.24.3
tensorflow==2.13.0
Pillow==10.0.0
retina-face==0.0.13

# Utilities
python-dateutil==2.8.2
pandas==2.3.0
joblib==1.5.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# CORS
fastapi-cors==0.0.6